Rename shared/router.rs to shared/route.rs and server/router.rs to server/route.rs.
