+++ server/src/routes/mod.rs
@@ -7,8 +7,8 @@ mod contact;
 pub use self::contact::*;
 mod not_found;
 pub use self::not_found::*;
-mod router;
-pub use router::*;
+mod route;
+pub use route::*;
 mod portfolio;
 pub use self::portfolio::*;
 mod flower;
